backend/algorithm/graph_pair/requirements.txt
imgaug==0.4.0
opencv_contrib_python_headless==4.4.0.46
PyMySQL==1.0.2
requests==2.22.0
six==1.14.0
lmdb==1.0.0
Flask==1.1.1
pyclipper==1.2.1
tqdm==4.42.1
numpy==1.19.0
Shapely==1.7.1
lanms==1.0.2
paddle==1.0.2
paddlehub==1.8.3
paddleslim==1.2.0
Pillow==8.1.0
python_Levenshtein==0.12.1
PyYAML==5.4.1
sip==6.0.0
visualdl==2.1.0
jieba_fast
Pillow